requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.1
pydantic>=2.4.2
pydantic[email]>=2.4.2
pydantic-settings>=2.0.0
slack-sdk>=3.23.0
slack-bolt>=1.18.0
PyGithub>=2.1.1
supabase>=2.0.3
pytest>=7.4.3
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
aiohttp>=3.8.6
websockets>=12.0
apscheduler>=3.10.4
python-multipart>=0.0.6
cryptography>=41.0.5
PyJWT>=2.8.0
